requirements.txt
pandas==1.5.3
requests==2.28.1
google-generativeai==0.1.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.26.0
fastapi==0.68.1
uvicorn==0.30.3